pyproject.toml
[tool.poetry]
name = "data-api"
version = "0.1.0"
description = ""
authors = ["Our World In Data <tech@worldindata.org>"]

[tool.poetry.dependencies]
python = "^3.9"
owid-catalog = {path = "vendor/owid-catalog-py", develop = true}
fastapi = "^0.67.0"
uvicorn = "^0.17.6"
pydantic = {extras = ["dotenv"], version = "^1.9.1"}
pandas = "^1.4.2"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.39"}
mysqlclient = "^2.1.0"
rich = "^12.4.4"
typer = "^0.4.1"
duckdb = "^0.4.0"
duckdb-engine = "^0.1.11"
structlog = "^21.5.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^2.10.1"
autoflake = "^1.4"
flake8 = "^3.8.4"
mypy = "^0.961"
isort = "^5.0"
pre-commit = "^2.8.2"
black = {version = "^22.3.0", extras = ["jupyter"]}
ipykernel = "^6.13.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
